+++ demisto_sdk/commands/prepare_content/preparers/marketplace_incident_to_alert_playbooks_prepare.py
@@ -1,5 +1,5 @@
 import logging
-import re
+from demisto_sdk.commands.prepare_content.preparers.incident_to_alert import prepare_descriptions_and_names
 
 from demisto_sdk.commands.common.constants import MarketplaceVersions
 
@@ -24,78 +24,8 @@ class MarketplaceIncidentToAlertPlaybooksPreparer:
 
         """
 
-        MarketplaceIncidentToAlertPlaybooksPreparer.prepare_descriptions_and_names(
+        prepare_descriptions_and_names(
             data, marketplace
         )
 
         return data
-
-    @staticmethod
-    def prepare_descriptions_and_names(data: dict, marketplace: MarketplaceVersions):
-        # Replace incidents to alerts only for XSIAM
-        replace_incident_to_alert = marketplace == MarketplaceVersions.MarketplaceV2
-
-        for k, v in data.get("tasks", {}).items():
-            if description := v.get("task", {}).get("description", ""):
-                if replace_incident_to_alert:
-                    description = MarketplaceIncidentToAlertPlaybooksPreparer.replace_incident_to_alert(description)
-
-                MarketplaceIncidentToAlertPlaybooksPreparer.remove_wrapper_from_incident(
-                    data, "description", description, k
-                )
-
-            if name := v.get("task", {}).get("name", ""):
-                if replace_incident_to_alert:
-                    name = MarketplaceIncidentToAlertPlaybooksPreparer.replace_incident_to_alert(name)
-
-                MarketplaceIncidentToAlertPlaybooksPreparer.remove_wrapper_from_incident(data, "name", name, k)
-
-        if description := data.get("description"):
-            if replace_incident_to_alert:
-                description = MarketplaceIncidentToAlertPlaybooksPreparer.replace_incident_to_alert(description)
-
-            MarketplaceIncidentToAlertPlaybooksPreparer.remove_wrapper_from_incident(data, "description", description)
-
-        if name := data.get("name"):
-            if replace_incident_to_alert:
-                name = MarketplaceIncidentToAlertPlaybooksPreparer.replace_incident_to_alert(name)
-
-            MarketplaceIncidentToAlertPlaybooksPreparer.remove_wrapper_from_incident(data, "name", name)
-
-    @staticmethod
-    def replace_incident_to_alert(name_or_description_field_content: str) -> str:
-        replacements = {r"(?<!<-)incident(?!->)": "alert",
-                   r"(?<!<-)Incident(?!->)": "Alert",
-                   r"(?<!<-)incidents(?!->)": "alerts",
-                   r"(?<!<-)Incidents(?!->)": "Alerts"}
-
-        new_content = name_or_description_field_content
-
-        for pattern, replace_with in replacements.items():
-            new_content = re.sub(
-                pattern, replace_with, new_content
-            )
-
-        return new_content
-
-    @staticmethod
-    def remove_wrapper_from_incident(
-        data: dict,
-        name_or_description_field: str,
-        name_or_description_field_content: str,
-        task_key: str = None,
-    ):
-        replacements = {r"<-incident->": "incident",
-                   r"<-Incident->": "Incident",
-                   r"<-incidents->": "incidents",
-                   r"<-Incidents->": "Incidents"}
-        new_content = name_or_description_field_content
-        for pattern, replace_with in replacements.items():
-            new_content = re.sub(
-                pattern, replace_with, new_content
-            )
-
-        if task_key:
-            data["tasks"][task_key]["task"][name_or_description_field] = new_content
-        else:
-            data[name_or_description_field] = new_content
